fix: aarch64 native porting fixes for pfr_play standalone build
- region_map.c: NULL guards for sMapCursor (SIGSEGV at fault_addr=0x2)
- scrcmd.c: ScrCmd_applymovement/vmessage use ScriptReadPtr not ScriptReadWord
- script_movement.c: defensive NULL check for movementScript pointer
- event_data.h: declare gSpecialVar_0x8014 (was defined but not declared)
- field_effect.c: use HostPtrStore for vblankCallback (64-bit safe)
- new_menu_helpers.c: (uintptr_t)ptr instead of (u32)ptr for heap pointers
- Various HOST_NATIVE guards for 64-bit pointer safety

+++ src/hall_of_fame.c
@@ -1122,7 +1122,7 @@ static void HallOfFame_PrintPlayerInfo(u8 unused1, u8 unused2)
 
 static void ClearVramOamPltt_LoadHofPal(void)
 {
-    u32 vramOffset, oamOffset, plttOffset;
+    uintptr_t vramOffset, oamOffset, plttOffset;
     u32 vramSize, oamSize, plttSize;
 
     vramOffset = (VRAM);
