+++ src/shop.c
@@ -273,14 +273,14 @@ static void Task_ShopMenu(u8 taskId)
 
 static void Task_HandleShopMenuBuy(u8 taskId)
 {
-    SetWordTaskArg(taskId, 0xE, (u32)CB2_InitBuyMenu);
+    SetWordTaskArg(taskId, 0xE, (uintptr_t)CB2_InitBuyMenu);
     FadeScreen(FADE_TO_BLACK, 0);
     gTasks[taskId].func = Task_GoToBuyOrSellMenu;
 }
 
 static void Task_HandleShopMenuSell(u8 taskId)
 {
-    SetWordTaskArg(taskId, 0xE, (u32)CB2_GoToSellMenu);
+    SetWordTaskArg(taskId, 0xE, (uintptr_t)CB2_GoToSellMenu);
     FadeScreen(FADE_TO_BLACK, 0);
     gTasks[taskId].func = Task_GoToBuyOrSellMenu;
 }
